backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datetime>=5.5",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "google>=3.0.0",
    "google-api-python-client>=2.187.0",
    "google-auth>=2.43.0",
    "google-auth-oauthlib>=1.2.2",
    "joblib>=1.5.2",
    "numpy>=2.3.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "typing>=3.10.0.0",
    "xgboost>=3.1.2",
]
